requirements.txt
hikari[speedups]
hikari-tanjun
uvloop
python-dotenv
